MainForm cleaned with BookinConstraints, and minor changes
return type of getCost changed to double, added comments in DAL, renamed FinalPageForm to FinalPagePanel, filled in method to set all fields

+++ src/internal/HotelDesc.java
@@ -155,7 +155,7 @@ public class HotelDesc {
     
     
     //TODO: COMPLETE METHOD! DUMMY BODY
-    public double getCost(String type, int rooms, LocalDate checkIn, LocalDate checkOut){
+    public int getCost(String type, int rooms, LocalDate checkIn, LocalDate checkOut){
         return 3450;
     }
